refactor: move security test to tests/test_cloud.py and remove separate test file
- Removed `tests/test_auth_security.py`.
- Moved `test_token_file_permissions` to `tests/test_cloud.py`.
- Refactored `tests/test_cloud.py` to use per-test skipping logic instead of module-level skip, allowing unit tests like `test_token_file_permissions` to run without `TC_CLOUD_TEST` env var.
